refactor: keyword -> tmi 용어 변경
