+++ IntroVerse/Presentation/SeokHwan/SeokHwanViewController.swift
@@ -9,7 +9,7 @@ final class SeokHwanViewController: UIViewController {
     private lazy var profileImageView = SeokHwanProfileImageView()
     private lazy var headerView = SeokHwanHeaderView()
     private lazy var dividerView = SeokHwanDividerView()
-    private lazy var keywordsStackView = SeokHwanKeywordsStackView()
+    private lazy var tmiStackView = SeokHwanTMIStackView()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -31,7 +31,7 @@ private extension SeokHwanViewController {
         contentView.addSubview(profileImageView)
         contentView.addSubview(headerView)
         contentView.addSubview(dividerView)
-        contentView.addSubview(keywordsStackView)
+        contentView.addSubview(tmiStackView)
     }
 
     func configureDelegates() {
@@ -60,7 +60,7 @@ private extension SeokHwanViewController {
             make.width.equalToSuperview().inset(20)
             make.centerX.equalToSuperview()
         }
-        keywordsStackView.snp.makeConstraints { make in
+        tmiStackView.snp.makeConstraints { make in
             make.top.equalTo(dividerView.snp.bottom).offset(20)
             make.width.equalToSuperview().inset(20)
             make.centerX.equalToSuperview()
